Task3 Версия 2

+++ pestrikov.id/Task3/Menu.cpp
@@ -5,7 +5,7 @@ Menu::Menu() {
 	selected = 0;
 }
 
-Menu::Menu(size_t count) : Menu() {
+Menu::Menu(int count) : Menu() {
 	commands.resize(count);
 }
